requirements.txt
requests
beautifulsoup4
pandas
transformers
torch
newspaper3k
nltk
matplotlib
gradio
python-dotenv